+++ meetup2apricot/event_processor.py
@@ -4,10 +4,6 @@ events and photos already seen."""
 
 from . import dryrun
 from .meetup_to_apricot_event_adaptor import MeetupToApricotEventAdaptor
-from .event_registration_type import (
-    make_apricot_registration_type,
-    make_meetup_registration_type,
-)
 import pickle
 import logging
 
@@ -24,6 +20,7 @@ class EventProcessor:
         latest_start_time,
         known_events,
         photo_cache,
+        event_registration_type_maker,
         apricot_api,
         cache_path,
         event_tagger,
@@ -31,12 +28,14 @@ class EventProcessor:
     ):
         """Initialize with the earliest and latest event start times, a
         dictionary of previously processed known events (indexed by Meetup
-        event ID), a photo cache; a Wild Apricot API interface, a path to the
-        cache file; an event tagger, and a dry run flag."""
+        event ID), a photo cache, an event registration type maker, a Wild
+        Apricot API interface, a path to the cache file; an event tagger, and a
+        dry run flag."""
         self.earliest_start_time = earliest_start_time
         self.latest_start_time = latest_start_time
         self.known_events = known_events
         self.photo_cache = photo_cache
+        self.event_registration_type_maker = event_registration_type_maker
         self.apricot_api = apricot_api
         self.cache_path = cache_path
         self.event_tagger = event_tagger
@@ -87,32 +86,71 @@ class EventProcessor:
         return apricot_event_id
 
     def add_event_registration_types(self, meetup_event, apricot_event_id):
+        """Add event registration types for a Wild Apricot event based on a
+        Meetup event."""
+        reg_types = self.gather_registration_types(meetup_event, apricot_event_id)
+        for reg_type in reg_types:
+            self.apricot_api.add_registration_type(reg_type.for_json())
+            self.log_add_event_registration_type(apricot_event_id, reg_type)
+
+    def gather_registration_types(self, meetup_event, apricot_event_id):
+        """Gather event registration types for a Wild Apricot event based on a
+        Meetup event."""
         meetup_count = meetup_event.yes_rsvp_count
         if meetup_event.rsvp_limit:
             apricot_count = meetup_event.rsvp_limit - meetup_count
         else:
             apricot_count = None
-        for reg_type in [
-            make_meetup_registration_type(
-                apricot_event_id, meetup_event.yes_rsvp_count
-            ),
-            make_apricot_registration_type(
+        meetup_type = self.make_meetup_type(
+            apricot_event_id, meetup_event.yes_rsvp_count
+        )
+        if meetup_event.members_only:
+            apricot_type = self.make_members_only_type(
                 apricot_event_id, apricot_count, meetup_event.fee_amount
-            ),
-        ]:
-            self.apricot_api.add_registration_type(reg_type.for_json())
-            if reg_type.maximum_registrants_count is None:
-                display_count = "unlimited"
-            else:
-                display_count = f"{reg_type.maximum_registrants_count:d}"
-            self.logger.info(
-                "add_event_registration_types: apricot_id=%d "
-                "maximum_registrants_count=%s price=%.2f name=%r",
-                apricot_event_id,
-                display_count,
-                reg_type.price,
-                reg_type.name,
             )
+
+        else:
+            apricot_type = self.make_rsvp_type(
+                apricot_event_id, apricot_count, meetup_event.fee_amount
+            )
+        return [meetup_type, apricot_type]
+
+    def make_meetup_type(self, apricot_event_id, count):
+        """Make a Meetup event registration type given a Wild Apricot event ID
+        and count of Meetup registrants."""
+        return self.event_registration_type_maker.make_meetup_registration_type(
+            apricot_event_id, count
+        )
+
+    def make_rsvp_type(self, apricot_event_id, count, fee):
+        """Make an RSVP event registration type given a Wild Apricot event ID,
+        a count of registrations allowed on Wild Apricot, and a fee."""
+        return self.event_registration_type_maker.make_apricot_registration_type(
+            apricot_event_id, count, fee
+        )
+
+    def make_members_only_type(self, apricot_event_id, count, fee):
+        """Make an members only event registration type given a Wild Apricot
+        event ID, a count of registrations allowed on Wild Apricot, and a
+        fee."""
+        return self.event_registration_type_maker.make_members_only_registration_type(
+            apricot_event_id, count, fee
+        )
+
+    def log_add_event_registration_type(self, apricot_event_id, reg_type):
+        """Log adding an event registration type for a Wild Apricot event."""
+        if reg_type.maximum_registrants_count is None:
+            display_count = "unlimited"
+        else:
+            display_count = f"{reg_type.maximum_registrants_count:d}"
+        self.logger.info(
+            "add_event_registration_types: apricot_id=%d "
+            "maximum_registrants_count=%s price=%.2f name=%r",
+            apricot_event_id,
+            display_count,
+            reg_type.price,
+            reg_type.name,
+        )
 
     def record_event(self, meetup_event, apricot_event_id):
         """Record the known event to ignore in the future."""