+++ HISTORY.rst
@@ -15,6 +15,12 @@ Recent History
 .. Next Release
 .. ------------------
 
+1.2.2 (2020-11-20)
+------------------
+
+* For "members only" events, restrict registrations instead of event visibility.
+* Document how to run meetup2apricot.
+
 1.2.1 (2020-11-18)
 ------------------